feat(blueprints): extract APIs to blueprint files
