+++ bruno/collection.bru
@@ -1,0 +1,20 @@
+auth {
+  mode: oauth2
+}
+
+auth:oauth2 {
+  grant_type: password
+  access_token_url: {{endpoint}}/oauth/token
+  refresh_token_url: 
+  username: jane.doe@mail.com
+  password: {{admin-pass}}
+  client_id: 
+  client_secret: 
+  scope: 
+  credentials_placement: body
+  credentials_id: Authorization
+  token_placement: header
+  token_header_prefix: Bearer
+  auto_fetch_token: true
+  auto_refresh_token: false
+}
