+++ bruno/profile/get profile.bru
@@ -9,7 +9,3 @@ get {
   body: none
   auth: inherit
 }
-
-headers {
-  Authorization: Bearer {{admin-token}}
-}
