+++ src/fit/blueprints/__init__.py
@@ -1,0 +1,5 @@
+from .user_blueprint import user_bp
+from .auth_blueprint import auth_bp
+from .coach_blueprint import coach_bp
+
+__all__ = ['user_bp', 'auth_bp', 'coach_bp']
